[autofix] Stabilize content version lifecycle test
